Move the hook to a subdir
